@@ -1,6 +1,6 @@
 {
   "name": "angular-starter",
-  "version": "12.2.12",
+  "version": "13.1.1",
   "scripts": {
     "ng": "ng",
     "start": "ng serve --port 4200",
@@ -15,42 +15,43 @@
   "author": {
     "name": "danny",
     "email": "danny@ganatan.com",
-    "url": "http://www.ganatan.com/"
+    "url": "https://www.ganatan.com/"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "12.2.12",
-    "@angular/common": "12.2.12",
-    "@angular/compiler": "12.2.12",
-    "@angular/core": "12.2.12",
-    "@angular/forms": "12.2.12",
-    "@angular/platform-browser": "12.2.12",
-    "@angular/platform-browser-dynamic": "12.2.12",
-    "@angular/platform-server": "12.2.12",
-    "@angular/router": "12.2.12",
-    "@angular/service-worker": "12.2.12",
+    "@angular/animations": "13.1.1",
+    "@angular/common": "13.1.1",
+    "@angular/compiler": "13.1.1",
+    "@angular/core": "13.1.1",
+    "@angular/forms": "13.1.1",
+    "@angular/platform-browser": "13.1.1",
+    "@angular/platform-browser-dynamic": "13.1.1",
+    "@angular/platform-server": "13.1.1",
+    "@angular/router": "13.1.1",
+    "@angular/service-worker": "13.1.1",
     "@fortawesome/fontawesome-free": "5.15.4",
-    "@nguniversal/express-engine": "12.1.3",
+    "@nguniversal/express-engine": "13.0.1",
     "bootstrap": "5.1.3",
-    "express": "4.17.1",
+    "express": "4.17.2",
+    "prismjs": "1.25.0",
     "rxjs": "7.4.0",
     "tslib": "2.3.1",
     "zone.js": "0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "12.2.12",
-    "@angular/cli": "12.2.12",
-    "@angular/compiler-cli": "12.2.12",
-    "@nguniversal/builders": "12.1.3",
+    "@angular-devkit/build-angular": "13.1.2",
+    "@angular/cli": "13.1.2",
+    "@angular/compiler-cli": "13.1.1",
+    "@nguniversal/builders": "13.0.1",
     "@types/express": "4.17.13",
-    "@types/jasmine": "3.10.1",
-    "@types/node": "16.11.6",
+    "@types/jasmine": "3.10.2",
+    "@types/node": "17.0.0",
     "jasmine-core": "3.10.1",
-    "karma": "6.3.6",
+    "karma": "6.3.9",
     "karma-chrome-launcher": "3.1.0",
-    "karma-coverage": "2.0.3",
+    "karma-coverage": "2.1.0",
     "karma-jasmine": "4.0.1",
     "karma-jasmine-html-reporter": "1.7.0",
-    "typescript": "4.3.5"
+    "typescript": "4.5.4"
   }
 }
