@@ -7,13 +7,58 @@ import { NotFoundComponent } from './modules/general/not-found/not-found.compone
 const routes: Routes = [
   { path: '', component: HomeComponent, },
   {
-    path: 'httpclient',
-    loadChildren: () => import('./modules/application/items/items.module')
-      .then(mod => mod.ItemsModule)
-  },  
+    path: 'crud',
+    loadChildren: () => import('./modules/application/example-crud/tutorial.module')
+      .then(mod => mod.TutorialModule)
+  },
   {
-    path: 'template-driven-forms',
-    loadChildren: () => import('./modules/application/example-template-driven-forms/tutorial.module')
+    path: 'news',
+    loadChildren: () => import('./modules/application/news/news.module')
+      .then(mod => mod.NewsModule)
+  },
+  {
+    path: 'news/:id',
+    loadChildren: () => import('./modules/application/news-form/news-form.module')
+      .then(mod => mod.NewsFormModule)
+  },
+  {
+    path: 'boxoffice',
+    loadChildren: () => import('./modules/application/boxoffice/boxoffice.module')
+      .then(mod => mod.BoxofficeModule)
+  },
+  {
+    path: 'boxoffice/:id',
+    loadChildren: () => import('./modules/application/boxoffice-form/boxoffice-form.module')
+      .then(mod => mod.BoxofficeFormModule)
+  },
+  {
+    path: 'modal',
+    loadChildren: () => import('./modules/application/example-modal/tutorial.module')
+      .then(mod => mod.TutorialModule)
+  },
+  {
+    path: 'prism',
+    loadChildren: () => import('./modules/application/example-prism/tutorial.module')
+      .then(mod => mod.TutorialModule)
+  },
+  {
+    path: 'angular',
+    loadChildren: () => import('./modules/application/angular/angular.module')
+      .then(mod => mod.AngularModule)
+  },
+  {
+    path: 'httpclient',
+    loadChildren: () => import('./modules/application/example-httpclient/items.module')
+      .then(mod => mod.ItemsModule)
+  },
+  {
+    path: 'services',
+    loadChildren: () => import('./modules/application/example-services/tutorial.module')
+      .then(mod => mod.TutorialModule)
+  },
+  {
+    path: 'forms',
+    loadChildren: () => import('./modules/application/example-forms/tutorial.module')
       .then(mod => mod.TutorialModule)
   },
   {
@@ -22,20 +67,20 @@ const routes: Routes = [
       .then(mod => mod.TutorialModule)
   },
   {
-    path: 'services',
-    loadChildren: () => import('./modules/application/example-services/tutorial.module')
-      .then(mod => mod.TutorialModule)
-  },
-  {
-    path: 'reactive-form',
-    loadChildren: () => import('./modules/application/example-reactive-form/tutorial.module')
-      .then(mod => mod.TutorialModule)
-  },    
-  {
     path: 'bootstrap',
     loadChildren: () => import('./modules/application/example-bootstrap/tutorial.module')
       .then(mod => mod.TutorialModule)
-  },  
+  },
+  {
+    path: 'signup',
+    loadChildren: () => import('./modules/general/signup/signup.module')
+      .then(mod => mod.SignupModule)
+  },
+  {
+    path: 'login',
+    loadChildren: () => import('./modules/general/login/login.module')
+      .then(mod => mod.LoginModule)
+  },
   {
     path: 'contact',
     loadChildren: () => import('./modules/general/contact/contact.module')
@@ -46,18 +91,14 @@ const routes: Routes = [
     loadChildren: () => import('./modules/general/about/about.module')
       .then(mod => mod.AboutModule)
   },
-  {
-    path: 'signin',
-    loadChildren: () => import('./modules/general/signin/signin.module')
-      .then(mod => mod.SigninModule)
-  },
   { path: '**', component: NotFoundComponent }
 ];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes, {
-    initialNavigation: 'enabled'
-})],
+    initialNavigation: 'enabled',
+    scrollPositionRestoration: 'enabled'
+  })],
   exports: [RouterModule],
   declarations: []
 })
