@@ -1,19 +1,26 @@
 export const ITEMS: any[] =
   [
-    { id: 1, name: 'Alerts', link: 'alerts' },
-    { id: 2, name: 'Badge', link: 'badge' },
-    { id: 3, name: 'Blockquotes', link: 'blockquotes' },
-    { id: 4, name: 'Breadcrumb', link: 'breadcrumb' },
-    { id: 5, name: 'Buttons', link: 'buttons' },
-    { id: 6, name: 'Collapse', link: 'collapse' },
-    { id: 7, name: 'Dropdowns', link: 'dropdowns' },
-    { id: 8, name: 'Forms', link: 'forms' },
-    { id: 9, name: 'List Group', link: 'list-group' },
-    { id: 10, name: 'Modal', link: 'modal' },
-    { id: 11, name: 'Pagination', link: 'pagination' },
-    { id: 12, name: 'Popovers', link: 'popovers' },
-    { id: 13, name: 'Progress', link: 'progress' },
-    { id: 14, name: 'Spinners', link: 'spinners' },
-    { id: 15, name: 'Toasts', link: 'toasts' },
-    { id: 16, name: 'Tooltips', link: 'tooltips' },
+    { id: 1, name: 'Accordions', link: 'accordions', icon: 'far fa-address-card' },
+    { id: 2, name: 'Alerts', link: 'alerts', icon: 'fas fa-user' },
+    { id: 3, name: 'Badges', link: 'badges', icon: 'fas fa-user-friends' },
+    { id: 4, name: 'Blockquotes', link: 'blockquotes', icon: 'fas fa-house-user' },
+    { id: 5, name: 'Breadcrumb', link: 'breadcrumb', icon: 'fab fa-app-store' },
+    { id: 6, name: 'Buttons', link: 'buttons', icon: 'fab fa-artstation' },
+    { id: 7, name: 'Cards', link: 'cards', icon: 'fab fa-asymmetrik' },
+    { id: 8, name: 'Checkbox', link: 'checkbox', icon: 'fas fa-atom' },
+    { id: 9, name: 'Collapse', link: 'collapse', icon: 'fas fa-balance-scale-right' },
+    { id: 10, name: 'Dropdowns', link: 'dropdowns', icon: 'fas fa-bahai' },
+    { id: 11, name: 'Forms', link: 'forms', icon: 'fas fa-basketball-ball' },
+    { id: 12, name: 'Images', link: 'images', icon: 'fab fa-battle-net' },
+    { id: 13, name: 'Modal', link: 'modal', icon: 'fab fa-canadian-maple-leaf' },
+    { id: 14, name: 'Pagination', link: 'pagination', icon: 'far fa-address-card' },
+    { id: 15, name: 'Radio', link: 'radio', icon: 'fas fa-cart-plus' },
+    { id: 16, name: 'Select', link: 'select', icon: 'fab fa-centos' },
+    { id: 17, name: 'Tables', link: 'tables', icon: 'fas fa-chart-line' },
+    { id: 18, name: 'Toasts', link: 'toasts', icon: 'fas fa-cogs' },
+    { id: 19, name: 'Tooltips', link: 'tooltips', icon: 'fab fa-edge' },
+    { id: 20, name: 'Popovers', link: 'popovers', icon: 'fab fa-jedi-order' },
+    { id: 21, name: 'Progress', link: 'progress', icon: 'fab fa-galactic-republic' },
+    { id: 22, name: 'Spinners', link: 'spinners', icon: 'fab fa-empire' },
+    { id: 23, name: 'Typography', link: 'typography', icon: 'fas fa-dove' },
   ];
