@@ -3,7 +3,7 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { TutorialComponent } from './tutorial.component';
 import { SafePipe } from './safe.pipe';
 
-describe('TutorialComponent', () => {
+describe('ExampleServicesComponent', () => {
   let component: TutorialComponent;
   let fixture: ComponentFixture<TutorialComponent>;
 
