@@ -1,4 +1,11 @@
-import { Component } from '@angular/core';
+import { Component, Inject } from '@angular/core';
+import { PLATFORM_ID } from '@angular/core';
+import { Meta, Title } from '@angular/platform-browser';
+import { isPlatformBrowser } from '@angular/common';
+
+import { Quote } from './quote';
+
+import { Observable, of } from 'rxjs';
 
 @Component({
   selector: 'app-about',
@@ -9,20 +16,28 @@ export class AboutComponent {
 
   dependencies: any;
   features: any;
+  quote: Quote;
+  id: number;
 
-  constructor() {
+  constructor(
+    @Inject(PLATFORM_ID) private platformId: object,
+    private meta: Meta,
+    private titleService: Title) {
+
+    this.id = 0;
+    this.quote = new Quote();
     this.dependencies = {
       frontend: [
-        { name: 'Angular 12.2.12' },
-        { name: 'Angular CLI 12.2.12' },
-        { name: 'Angular Universal 12.1.3' },
-        { name: 'Font Awesome 5.15.4' },
+        { name: 'Angular 13.1.1' },
+        { name: 'Angular CLI 13.1.2' },
+        { name: 'Angular Universal 13.0.1' },
         { name: 'Bootstrap 5.1.3' },
+        { name: 'Font Awesome 5.15.4' },
       ],
       backend: [
-        { name: 'Node.js' },
-        { name: 'Express' },
-        { name: 'pg-promise' },
+        { name: 'Node.js 14.17.3' },
+        { name: 'Express 4.17.1' },
+        { name: 'pg-promise 10.10.2' },
       ]
     };
 
@@ -30,63 +45,58 @@ export class AboutComponent {
       frontend: [
         {
           name: 'Angular CLI',
-          englishTutorial: 'http://www.ganatan.com/tutorials/getting-started-with-angular',
-          frenchTutorial: 'http://www.ganatan.com/tutorials/demarrer-avec-angular',
+          englishTutorial: 'https://www.ganatan.com/tutorials/getting-started-with-angular',
+          frenchTutorial: 'https://www.ganatan.com/tutorials/demarrer-avec-angular',
         },
         {
           name: 'Routing',
-          englishTutorial: 'http://www.ganatan.com/tutorials/routing-with-angular',
-          frenchTutorial: 'http://www.ganatan.com/tutorials/routing-avec-angular',
+          englishTutorial: 'https://www.ganatan.com/tutorials/routing-with-angular',
+          frenchTutorial: 'https://www.ganatan.com/tutorials/routing-avec-angular',
         },
         {
           name: 'Lazy loading',
-          englishTutorial: 'http://www.ganatan.com/tutorials/lazy-loading-with-angular',
-          frenchTutorial: 'http://www.ganatan.com/tutorials/lazy-loading-avec-angular',
+          englishTutorial: 'https://www.ganatan.com/tutorials/lazy-loading-with-angular',
+          frenchTutorial: 'https://www.ganatan.com/tutorials/lazy-loading-avec-angular',
         },
         {
           name: 'Bootstrap',
-          englishTutorial: 'http://www.ganatan.com/tutorials/bootstrap-with-angular',
-          frenchTutorial: 'http://www.ganatan.com/tutorials/bootstrap-avec-angular',
+          englishTutorial: 'https://www.ganatan.com/tutorials/bootstrap-with-angular',
+          frenchTutorial: 'https://www.ganatan.com/tutorials/bootstrap-avec-angular',
         },
         {
           name: 'Server side Rendering',
-          englishTutorial: 'http://www.ganatan.com/tutorials/server-side-rendering-with-angular-universal',
-          frenchTutorial: 'http://www.ganatan.com/tutorials/server-side-rendering-avec-angular-universal',
+          englishTutorial: 'https://www.ganatan.com/tutorials/server-side-rendering-with-angular-universal',
+          frenchTutorial: 'https://www.ganatan.com/tutorials/server-side-rendering-avec-angular-universal',
         },
         {
           name: 'HTTPClient',
-          englishTutorial: 'http://www.ganatan.com/tutorials/httpclient-with-angular',
-          frenchTutorial: 'http://www.ganatan.com/tutorials/httpclient-avec-angular',
+          englishTutorial: 'https://www.ganatan.com/tutorials/httpclient-with-angular',
+          frenchTutorial: 'https://www.ganatan.com/tutorials/httpclient-avec-angular',
         },
         {
           name: 'Transfer State',
-          englishTutorial: 'http://www.ganatan.com/tutorials/transfer-state-with-angular',
-          frenchTutorial: 'http://www.ganatan.com/tutorials/transfer-state-avec-angular',
+          englishTutorial: 'https://www.ganatan.com/tutorials/transfer-state-with-angular',
+          frenchTutorial: 'https://www.ganatan.com/tutorials/transfer-state-avec-angular',
         },
         {
           name: 'Progressive Web App',
-          englishTutorial: 'http://www.ganatan.com/tutorials/progressive-web-app-with-angular',
-          frenchTutorial: 'http://www.ganatan.com/tutorials/progressive-web-app-avec-angular',
+          englishTutorial: 'https://www.ganatan.com/tutorials/progressive-web-app-with-angular',
+          frenchTutorial: 'https://www.ganatan.com/tutorials/progressive-web-app-avec-angular',
         },
         {
           name: 'Search Engine optimization',
-          englishTutorial: 'http://www.ganatan.com/tutorials/search-engine-optimization-with-angular',
-          frenchTutorial: 'http://www.ganatan.com/tutorials/search-engine-optimization-avec-angular',
+          englishTutorial: 'https://www.ganatan.com/tutorials/search-engine-optimization-with-angular',
+          frenchTutorial: 'https://www.ganatan.com/tutorials/search-engine-optimization-avec-angular',
         },
         {
           name: 'Components',
           englishTutorial: null,
-          frenchTutorial: 'http://www.ganatan.com/tutorials/components-avec-angular',
+          frenchTutorial: 'https://www.ganatan.com/tutorials/components-avec-angular',
         },
         {
           name: 'Services',
           englishTutorial: null,
-          frenchTutorial: 'http://www.ganatan.com/tutorials/services-avec-angular',
-        },
-        {
-          name: 'Responsive Images List',
-          englishTutorial: null,
-          frenchTutorial: null,
+          frenchTutorial: 'https://www.ganatan.com/tutorials/services-avec-angular',
         },
       ],
       backend: [
@@ -99,6 +109,55 @@ export class AboutComponent {
       ]
     };
 
+  }
+
+  ngOnInit(): void {
+    if (isPlatformBrowser(this.platformId)) {
+      // MCP
+      // $('html,body').scrollTop(0);
+    }
+
+    this.loadQuote();
+
+    this.titleService.setTitle('angular.ganatan.com : A Propos');
+    this.meta.addTag({
+      name: 'author',
+      content: 'danny ganatan'
+    });
+    this.meta.updateTag(
+      {
+        name: 'description',
+        content: 'Cette application a été développée avec Angular version 13.1.1 et bootstrap 5.1.3' +
+          ' Elle applique le Routing, le Lazy loading, le Server side rendering et les Progressive Web App (PWA)'
+      });
+  }
+
+
+  loadQuote() {
+    let quotes = [
+      {
+        name: 'Lawrence of Arabia',
+        title: 'There is nothing in the desert and no man needs nothing',
+        link: 'https://en.wikipedia.org/wiki/Lawrence_of_Arabia_(film)'
+      },
+      {
+        name: 'Alien Prometheus',
+        title: 'Big things have small beginnings',
+        link: 'https://en.wikipedia.org/wiki/Prometheus_(2012_film)'
+      },
+      {
+        name: 'Blade Runner',
+        title: 'All those moments will be lost in time... like tears in rain... Time to die.',
+        link: 'https://en.wikipedia.org/wiki/Blade_Runner'
+      },
+    ];
+    let index = quotes.length;
+    let id = this.id;
+    while (this.id === id) {
+      id = Math.floor(Math.random() * index);
+    }
+    this.id = id;
+    this.quote = quotes[id];
   }
 
 }
