@@ -1,8 +1,9 @@
 import { Component, OnInit } from '@angular/core';
-import { environment } from '../../../../environments/environment';
 import { Inject, PLATFORM_ID } from '@angular/core';
 import { isPlatformBrowser } from '@angular/common';
 import { Meta, Title } from '@angular/platform-browser';
+
+import { environment } from '../../../../environments/environment';
 
 @Component({
   selector: 'app-home',
@@ -22,53 +23,22 @@ export class HomeComponent implements OnInit {
     @Inject(PLATFORM_ID) private platformId: object,
     private meta: Meta,
     private titleService: Title) {
-    this.features =
-      [
-        {
-          type: 'Bootstrap',
-          description: 'Bootstrap , Badges, Buttons, Collapse, Popovers, Toasts, Tooltips',
-          image: 'demo-bootstrap-prototype.png',
-          link: 'bootstrap'
-        },
-        {
-          type: 'Reactive Form',
-          description: 'A model-driven approach to handling form inputs',
-          image: 'demo-reactive-forms.png',
-          link: 'reactive-form'
-        },
-        {
-          type: 'Services',
-          description: 'Use services to view a playlist and a youtube player',
-          image: 'demo-services-playlist-youtube.png',
-          link: 'services'
-        },
-        {
-          type: 'Components',
-          description: 'Smartphone Component with Input, Output and Event Emitter',
-          image: 'demo-components.png',
-          link: 'components'
-        },
-        {
-          type: 'HttpClient',
-          description: 'Use an external API with the HttpClient module',
-          image: 'demo-httpclient.png',
-          link: 'httpclient'
-        },
-        {
-          type: 'Template Driven Forms',
-          description: 'Forms are the mainstay of business applications',
-          image: 'demo-template-driven-forms.png',
-          link: 'template-driven-forms'
-        },
-      ];
-
   }
 
   ngOnInit(): void {
 
-    this.loadScript('assets/params/js/index.js');
+    if (isPlatformBrowser(this.platformId)) {
+      let navMain = document.getElementById('navbarCollapse');
+      if (navMain) {
+        navMain.onclick = function () {
+          if (navMain) {
+            navMain.classList.remove("show");
+          }
+        }
+      }
+    }
 
-    this.titleService.setTitle('angular.ganatan: Une Application Web avec Angular');
+    this.titleService.setTitle('angular.ganatan.com : An Angular Application Web');
     this.meta.addTag({
       name: 'author',
       content: 'danny ganatan'
@@ -76,13 +46,12 @@ export class HomeComponent implements OnInit {
     this.meta.updateTag(
       {
         name: 'description',
-        content: 'Cette application a été développée avec Angular version 12.2.12 et bootstrap 5.1.3' +
-          ' Elle applique le Routing, le Lazy loading, le Server side rendering et les Progressive Web App (PWA)'
+        content: 'This application was developed with Angular version 13.1.1 and bootstrap 5.1.3' +
+          ' It applies Routing, Lazy loading, Server side rendering and Progressive Web App (PWA)'
       });
   }
 
   loadScript(name: string): void {
-
     if (isPlatformBrowser(this.platformId)) {
       const s = document.createElement('script');
       s.type = 'text/javascript';
@@ -91,7 +60,6 @@ export class HomeComponent implements OnInit {
       document.getElementsByTagName('head')[0].appendChild(s);
     }
   }
-
 
 }
 
